move optimism tx meta to new file
